refactor(publish): remove placeholder.py file
Keep the dollar-year and purchaser-price adjustment local to the emission-factor table builder now that placeholder margin factors are gone.
